+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [#5](https://github.com/dblock/strava-ruby-client/issues/5): Added support for pagination - [@dblock](https://github.com/dblock).
 * Added `Strava::Api::Client#athlete_clubs` - [@dblock](https://github.com/dblock).
 * Added `Strava::Api::Client#club_activities` - [@dblock](https://github.com/dblock).
+* Added `Strava::Api::Client#create_activity` - [@dblock](https://github.com/dblock).
 * [#6](https://github.com/dblock/strava-ruby-client/issues/6): The `oauth-token` tool will handle the redirect in the browser - [@dblock](https://github.com/dblock).
 * Your contribution here.